Create cable_hash.py
Makes a hashtable for the cables. Unfortunately it loops through the first 36 houses instead of all 150 making it seem like it loops through 149 of them if you don't substract the output from the houses. Running the file gives a lot of information about which house has been connected and which one has not.